Move _Generating Jar_ build operation to `GeneratedGradleJarCache`
By moving the `BuildOperationExecutor` call to
`DefaultGeneratedGradleJarCache` so the generation of Gradle Kotlin
DSL API extensions Jar is also reported as a build operation.
